Split api/community.py into two files to avoid circular import error.
