+++ pylemmy/api/person.py
@@ -1,7 +1,7 @@
 from typing import List, Optional
 
 from pylemmy.api.comment import CommentView
-from pylemmy.api.community import CommunityModeratorView
+from pylemmy.api.community2 import CommunityModeratorView
 from pylemmy.api.listing import SortType
 from pylemmy.api.post import PostView
 from pylemmy.api.utils import BaseApiModel
